backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
ecdsa==0.19.0
Flask==1.1.2
Flask-Cors==5.0.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
python-jose==3.3.0
rsa==4.9
six==1.17.0
SQLAlchemy==1.3.23
typing_extensions==4.12.2
Werkzeug==2.0.1
wrapt==1.17.2
